requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
email-validator
requests
playwright
python-jose[cryptography]
passlib[bcrypt]
slowapi
anyio
orjson
apscheduler
stripe
